refactor: make base runner class
